# Conflicts:
#	Stars-coffee-rest-api/CustomerDB.sqlite

+++ Stars-coffee-rest-api/src/main/java/org/feup/coffeeshop/model/entity/VoucherRequestEntity.java
@@ -39,7 +39,7 @@ public class VoucherRequestEntity {
     @Column
     private String description;
     @Column
-    private int value;
+    private Double value;
     @Column
     private String image;
     @Column
